chore(cascader): demo拆解与规范
